requirements.txt
openai
requests
pydub
python-dotenv 